Auth0/requirements.txt
Authlib==1.6.1
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
ecdsa==0.19.1
Flask==2.3.3
flask-cors==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
python-jose==3.5.0
requests==2.32.4
rsa==4.9.1
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
